Change RESTActions to RESTfulActions and place it into resource namespace
