+++ lib/jeckle/resource.rb
@@ -1,3 +1,5 @@
+require_relative 'resource/restful_actions'
+
 module Jeckle
   module Resource
     def self.included(base)
@@ -5,7 +7,7 @@ module Jeckle
 
       base.send :include, Jeckle::Model
       base.send :include, Jeckle::HTTP
-      base.send :include, Jeckle::RESTActions
+      base.send :include, Jeckle::Resource::RESTfulActions
     end
   end
 end
